Incomplete implementation of timezone setting on registration, my account...
Plug was pulled on the rewrite, so pushing this up unfinished.
Extra timezone fields added via a partial view, so fields exist on
registration, my account and timezone views.
Post actions on UserController incomplete or untested on UserController
for create, edit and timezoneChange.
At the moment, timezone is just a textbox. This needs to be implemented
as a Select pulling the data from DateTimeZone::listIdentifiers.
Need to add timezone support to all lesson related views and anywhere else
that a date and time are edited/displayed.